[CHG] refactored to logparser
