+++ logparser/kind.go
@@ -1,4 +1,4 @@
-package parser
+package logparser
 
 import (
 	"fmt"
