+++ logparser/kind_test.go
@@ -1,6 +1,6 @@
 // +build !integration
 
-package parser
+package logparser
 
 import (
 	"testing"
