+++ logparser/s3cloudfrontlogparser.go
@@ -1,4 +1,4 @@
-package parser
+package logparser
 
 import (
 	"regexp"
