+++ pipeline/sqs.go
@@ -5,18 +5,18 @@ import (
 
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/mpucholblasco/s3logsbeat/aws"
-	"github.com/mpucholblasco/s3logsbeat/parser"
+	"github.com/mpucholblasco/s3logsbeat/logparser"
 )
 
 // SQS SQS element to send thru pipeline
 type SQS struct {
 	*aws.SQS
-	logParser      parser.LogParser
+	logParser      logparser.LogParser
 	keyRegexFields *regexp.Regexp
 }
 
 // NewSQS creates a new SQS to be sent thru pipeline
-func NewSQS(session *session.Session, queueURL *string, logParser parser.LogParser, keyRegexFields *regexp.Regexp) *SQS {
+func NewSQS(session *session.Session, queueURL *string, logParser logparser.LogParser, keyRegexFields *regexp.Regexp) *SQS {
 	return &SQS{
 		SQS:            aws.NewSQS(session, queueURL),
 		logParser:      logParser,
